Created files parallelColorSearcher.py and singleParallelColorSearch.py
parallelColorSearcher.py imitates colorSearcher.py but introduces multiprocessing (not multi-threading due to its limitations in python). singleParallelColorSearch.py is exactly the same as singleColorSearch.py except that it uses the parallelColorSearcher instead of the regular one.

Note that this new functionallity has not yet been tested. I lack the external hard drive with the screenshots, else I would test it myself. You'll want to adjust the number of pooled processes and do runtime tests to find the optimal number of processes.